added splinesdrive class, fixed simplesplines references
