+++ commands/chassis/SimpleSplines.java
@@ -5,7 +5,7 @@ package org.usfirst.frc4904.standard.commands.chassis;
 import java.util.List;
 
 import org.usfirst.frc4904.robot.RobotMap;
-import org.usfirst.frc4904.standard.subsystems.chassis.SensorDrive;
+import org.usfirst.frc4904.standard.subsystems.chassis.SplinesDrive;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
@@ -22,7 +22,7 @@ import edu.wpi.first.wpilibj2.command.RamseteCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 
 public class SimpleSplines extends SequentialCommandGroup {
-  public SimpleSplines(SensorDrive robotDrive, Pose2d init_pos, List<Translation2d> inter_points, Pose2d final_pos, double maxVoltage, Command nextCommand){
+  public SimpleSplines(SplinesDrive robotDrive, Pose2d init_pos, List<Translation2d> inter_points, Pose2d final_pos, double maxVoltage, Command nextCommand){
     super(new RamseteCommand(
         TrajectoryGenerator.generateTrajectory(init_pos, inter_points, final_pos, new TrajectoryConfig(robotDrive.getAutoConstants().kMaxSpeedMetersPerSecond,
         robotDrive.getAutoConstants().kMaxAccelerationMetersPerSecondSquared)
@@ -46,15 +46,15 @@ public class SimpleSplines extends SequentialCommandGroup {
         robotDrive::tankDriveVolts, robotDrive), nextCommand);
   } 
 
-  public SimpleSplines(SensorDrive robotDrive, Pose2d init_pos, List<Translation2d> inter_points, Pose2d final_pos, double maxVoltage){
+  public SimpleSplines(SplinesDrive robotDrive, Pose2d init_pos, List<Translation2d> inter_points, Pose2d final_pos, double maxVoltage){
     this(robotDrive, init_pos, inter_points, final_pos, maxVoltage, new InstantCommand(() -> robotDrive.tankDriveVolts(0, 0)));
   }
 
-  public SimpleSplines(SensorDrive robotDrive, Pose2d init_pos, List<Translation2d> inter_points, Pose2d final_pos){
+  public SimpleSplines(SplinesDrive robotDrive, Pose2d init_pos, List<Translation2d> inter_points, Pose2d final_pos){
     this(robotDrive, init_pos, inter_points, final_pos, 10);
   }
 
-  public SimpleSplines(SensorDrive robotDrive, Pose2d init_pos, Pose2d final_pos){
+  public SimpleSplines(SplinesDrive robotDrive, Pose2d init_pos, Pose2d final_pos){
     this(robotDrive, init_pos, List.of(), final_pos);
   }
